railway-deployment/requirements.txt
Flask==3.0.0
Flask-Limiter==3.5.0
bcrypt==4.1.2
PyJWT==2.8.0
gunicorn==21.2.0
openpyxl==3.1.2
Werkzeug==3.0.1